requirements.txt
beautifulsoup4
feedparser
newspaper3k
pandas
python-dateutil
python-telegram-bot
requests
selenium
webdriver-manager
"lxml[html_clean]"
